requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
av==13.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
exceptiongroup==1.2.2
fire==0.7.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
numpy==2.1.2
openai==1.53.0
opencv-python-headless==4.10.0.84
picamera2==0.3.21
picamzero==1.0.1
pidng==4.0.9
piexif==1.1.3
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-prctl==1.8.1
requests==2.32.3
simplejpeg==1.7.6
sniffio==1.3.1
termcolor==2.5.0
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
v4l2-python3==0.3.5